requirements.txt
# Dota 2 AI Coach
flask>=3.0.0,<4
mss>=9.0.0
opencv-python-headless>=4.8.0,<5
numpy>=1.24.0,<3
PyQt6>=6.6.0,<7
openai>=1.12.0,<2
anthropic>=0.18.0,<1
httpx>=0.26.0,<1
pyyaml>=6.0.1,<7
python-dotenv>=1.0.0,<2
pynput>=1.7.0,<2
pyinstaller>=6.0.0,<7
